create_stocks
